+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0.0
 
+- Add support for fan speed monitoring
 - Add support for hashrate monitoring
 - Add version switch